test(laravel-bridge): sweep Laravel's statics in one base TestCase
Booting a provider leaves process-global state behind twice over: Gacela's
singletons, and Laravel's -- Artisan::starting() closures and the optimize
registries are static. Only one test class swept the Laravel half, so under
an unlucky seed another class's leaked bootstrappers replayed against an
application that had never bound the command classes, and InitCommand's
string argument turned the artisan constructor into an error. CI found the
seed; the sweeping now lives in a base TestCase every bridge test extends.
